🐛 fix: keep common onboarding tests focused
