+++ src/routes/onboarding/config.ts
@@ -1,4 +1,3 @@
-import { AGENT_ONBOARDING_ENABLED } from '@lobechat/business-const';
 import type { InterestAreaKey } from '@lobechat/const';
 import { INTEREST_AREA_KEYS } from '@lobechat/const';
 import type { LucideIcon } from 'lucide-react';
@@ -22,32 +21,12 @@ import {
   UsersIcon,
 } from 'lucide-react';
 
-export const ONBOARDING_AGENT_PATH = '/onboarding/agent';
-export const ONBOARDING_CLASSIC_PATH = '/onboarding/classic';
-
-export type OnboardingBranchPath = typeof ONBOARDING_AGENT_PATH | typeof ONBOARDING_CLASSIC_PATH;
-
-interface DeriveOnboardingBranchInput {
-  enableAgentOnboarding: boolean;
-  isDesktop: boolean;
-}
-
-/**
- * Decide which branch the user enters after the shared-prefix steps complete.
- * `AGENT_ONBOARDING_ENABLED` is the build-time master switch — when it is off,
- * the agent flow is unreachable regardless of the runtime feature flag.
- * Desktop and disabled-flag users also land on the classic flow; otherwise
- * the agent conversational flow is the default.
- */
-export const deriveOnboardingBranchPath = ({
-  enableAgentOnboarding,
-  isDesktop,
-}: DeriveOnboardingBranchInput): OnboardingBranchPath => {
-  if (!AGENT_ONBOARDING_ENABLED || isDesktop || !enableAgentOnboarding) {
-    return ONBOARDING_CLASSIC_PATH;
-  }
-  return ONBOARDING_AGENT_PATH;
-};
+export type { OnboardingBranchPath } from './branch';
+export {
+  deriveOnboardingBranchPath,
+  ONBOARDING_AGENT_PATH,
+  ONBOARDING_CLASSIC_PATH,
+} from './branch';
 
 /**
  * Predefined interest areas with icons and translation keys.
